created render-line-items.js to clean up code
