+++ cart/cart.js
@@ -1,26 +1,29 @@
 import { shirts } from '../data/shirts.js';
 import { cart } from '../data/cat-data.js';
 import { findById } from '../utils.js';
+import { renderLineItem } from '../render-line-items.js';
 // import { calcOrderTotal } from '../utils.js';
 
 const tbody = document.getElementById('table-body');
 for (let cartItem of cart){
     const shirtData = findById(cartItem.id, shirts);
-    const tr = document.createElement('tr');
+    // const tr = document.createElement('tr');
 
-    const tdName = document.createElement('td');
-    tdName.textContent = shirtData.name;
+    // const tdName = document.createElement('td');
+    // tdName.textContent = shirtData.name;
 
-    const tdPrice = document.createElement('td');
-    tdPrice.textContent = `$${shirtData.price}`;
+    // const tdPrice = document.createElement('td');
+    // tdPrice.textContent = `$${shirtData.price}`;
 
-    const tdQty = document.createElement('td');
-    tdQty.textContent = cartItem.qty;
+    // const tdQty = document.createElement('td');
+    // tdQty.textContent = cartItem.qty;
 
-    const tdTotal = document.createElement('td');
-    tdTotal.textContent = `$${cartItem.qty * shirtData.price}`;
+    // const tdTotal = document.createElement('td');
+    // tdTotal.textContent = `$${cartItem.qty * shirtData.price}`;
 
-    tr.append(tdName, tdPrice, tdQty, tdTotal);
+    const tr = renderLineItem(cartItem, shirtData);
+
+    // tr.append(tdName, tdPrice, tdQty, tdTotal);
     tbody.appendChild(tr);
 }
