+++ cart/index.html
@@ -21,12 +21,12 @@
             <th>Total</th>
         </thead>
         <tbody id="table-body">
-            <!-- <tr>
+            <tr>
                 <td>shirt 1</td>
                 <td>$100</td>
                 <td>6</td>
                 <td>$120</td>
-            </tr> -->
+            </tr>
         </tbody>
         <tfoot>
             <td></td>
