+++ data/cat-data.js
@@ -1,4 +1,4 @@
 export const cart = [
-    { id: '1', qty: 6 },
-    { id: '4', qty: 5 },
+    { id: '1', qty: 2 },
+    { id: '4', qty: 3 },
 ];
